Update worship time wording
For Easter (and some other services) an Event is created to announce
time changes. (For example, this year's Easter service is at 9:30am.)
This presents contradictory information with "This Sunday: <time>" as
it's currently presented. To compensate for that, we soften the wording
on the "This Sunday" section to make it a bit more obvious that the
hours aren't exactly authoritative, and can be overridden by an Event.